// app-ui/src/app/backend/model/csvFile.ts
/**
 * TIBCO Discover public API
 * TIBCO Discover public API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RedisFileInfo } from './redisFileInfo';


/**
 * The csv file info returned from backend and whether it is used by dataset.
 */
export interface CsvFile { 
    redisFileInfo?: RedisFileInfo;
    /**
     * Whether it is used by any dataset.
     */
    beingUsed?: boolean;
}