// app-ui/src/app/backend/model/redisFileInfo.ts
/**
 * TIBCO Discover public API
 * TIBCO Discover public API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * The copy of the backend redisFileInfo model. See https://discover.labs.tibcocloud.com/backswagger#/Files%20Operations/getRouteFileV2
 */
export interface RedisFileInfo { 
    ContentType: string;
    LastModified: string;
    OriginalFilename: string;
    OriginalEncoding: string;
    FileSize: string;
    newline: string;
    EscapeChar: string;
    QuoteChar: string;
    Separator: string;
    Encoding: string;
    OriginalNewLine: string;
    FileLocation: string;
}